api/requirements.txt
Django>=3.2.5,<4.0
djangorestframework>=3.14.0,<3.15
